% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_p_prior.R
\name{check_p_prior}
\alias{check_p_prior}
\alias{check_p_prior.numeric}
\alias{check_p_prior.matrix}
\alias{check_p_prior.xts}
\alias{check_p_prior.tbl}
\title{Functions to check the consistency of prior probabilities.}
\usage{
check_p_prior(p)

\method{check_p_prior}{numeric}(p)

\method{check_p_prior}{matrix}(p)

\method{check_p_prior}{xts}(p)

\method{check_p_prior}{tbl}(p)
}
\arguments{
\item{p}{A vector, a matrix, a xts or a tibble object.}
}
\value{
A matrix object with 1 column.
}
\description{
Functions to check the consistency of prior probabilities.
}
\examples{
#
}
\keyword{internal}
