\name{classification-methods}
\docType{methods}
\alias{classification-methods}
\alias{classification,matrix,numeric,missing-method}
\alias{classification,matrix,factor,missing-method}
\alias{classification,data.frame,missing,formula-method}
\alias{classification,ExpressionSet,character,missing-method}
\title{General method for classification with various methods}
\description{
  Perform classification for the following signatures:
}
\section{Methods}{
\describe{

\item{X = "matrix", y = "numeric", f = "missing"}{signature 1}

\item{X = "matrix", y = "factor", f = "missing"}{signature 2}

\item{X = "data.frame", y = "missing", f = "formula"}{signature 3}

\item{X = "ExpressionSet", y = "character", f = "missing"}{signature 4}
}
For further argument and output information, consult
\code{\link{classification}}.}
\keyword{multivariate}
