% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-cma.R
\name{cma-copula}
\alias{cma-copula}
\alias{new_copula}
\alias{obj_print_data.cma_copula}
\title{Internal vctrs methods}
\usage{
new_copula(x, ...)

\method{obj_print_data}{cma_copula}(x, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
No return value, called for side effects.
}
\description{
Internal vctrs methods
}
\keyword{internal}
