% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma-vctrs.R
\name{cma-marginal}
\alias{cma-marginal}
\alias{new_marginal}
\alias{obj_print_data.marginal}
\title{Internal vctrs methods}
\usage{
new_marginal(x, ...)

\method{obj_print_data}{marginal}(x, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
No return value, called for side effects.
}
\description{
Internal vctrs methods
}
\keyword{internal}
