% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-cma.R
\name{cma-panic}
\alias{cma-panic}
\alias{new_panic_copula}
\title{Internal vctrs methods}
\usage{
new_panic_copula(x, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
No return value, called for side effects.
}
\description{
Internal vctrs methods
}
\keyword{internal}
