% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMA.exported_functions.R
\name{CMA_codon}
\alias{CMA_codon}
\title{Correlated mutation analysis at the codon sequence level by multiple methods}
\usage{
CMA_codon(
  aa_alignment_file = system.file("extdata", "PF00014_full.fasta.test", package = "CMA"),
  aa_alignment_file_type = "fasta",
  codon_alignment_file = system.file("extdata", "PF00014_full.fasta.test.tranalignout",
    package = "CMA"),
  codon_alignment_file_type = "fasta",
  seqnameorindex = "Q7K3Y9_DROME/661-715"
)
}
\arguments{
\item{aa_alignment_file}{Input amino acid sequence alignment file name in a fasta or sth format}

\item{aa_alignment_file_type}{Amino acid alignment file format (fasta or sth)}

\item{codon_alignment_file}{Input codon sequence alignment file name in a fasta or sth format}

\item{codon_alignment_file_type}{Codon alignment file format (fasta or sth)}

\item{seqnameorindex}{Id or index of the sequence as given in the amino acid sequence alignment file
to evalute the correlated mutations for}
}
\value{
returns a data.frame of the measurements MI, OMES and DI at codon sequence level.
See below for a description of the measurements.
}
\description{
\code{CMA_codon} Calculates the Direct Coupling, OMES and MI measurments of all pairwise positions of
a protein sequence based on an input MSA file of codon level sequences in fasta or sth format
}
\details{
MI - **M**utual **I**nforamtion
calc MIp by:
1. Convert table MI data to a matrix.
2. Execute the algorithm of MIp (below)
3. Returns the same table with i & j and but the column, MIp

Algorithm (based on http://bioinformatics.oxfordjournals.org/content/24/3/333.full):
MI(a, mean(x)) = 1/m(Sum(MI(a, x)))
where n is the number of columns in the alignment, m = n-1 for convenience,
and the summation is over x=1 to n, x != a.

mean(MI) denotes the overall mean mutual information:
mean(MI) = 2/(m*n)(Sum(MI(x, y))),
where the indices run x=1 to m, y = x+1 to n.

The average product correction:
APC(a, b) = (MI(a, mean(x))*MI(b, mean(x)))/mean(MI)
Is an approximation to the background MI shared by positions a and b.
MIp denotes the difference between total observed MI and the APC:
MIp(a, b) = MI(a, b) - APC(a, b).
In order to avoid negative values, values are shifted by adding the minimum
}
\examples{
#Performing correlated mutatios annalysis at the codon level for Q7K3Y9_DROME/661-715 given
#PF00014 full alignment:
res.codon <- CMA_codon(aa_alignment_file = system.file("extdata",
                                                       "PF00014_full.fasta.test",
                                                       package ="CMA"),
                       aa_alignment_file_type = "fasta",
                       codon_alignment_file = system.file("extdata",
                                                          "PF00014_full.fasta.test.tranalignout",
                                                          package ="CMA"),
                       codon_alignment_file_type = "fasta",
                       seqnameorindex = "Q7K3Y9_DROME/661-715")
}
