% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_separation.R
\name{cma_separation}
\alias{cma_separation}
\alias{cma_separation.default}
\alias{cma_separation.matrix}
\alias{cma_separation.xts}
\alias{cma_separation.data.frame}
\alias{cma_separation.tbl}
\title{CMA Separation Step}
\usage{
cma_separation(x, p = NULL)

\method{cma_separation}{default}(x, p = NULL)

\method{cma_separation}{matrix}(x, p = NULL)

\method{cma_separation}{xts}(x, p = NULL)

\method{cma_separation}{data.frame}(x, p = NULL)

\method{cma_separation}{tbl}(x, p = NULL)
}
\arguments{
\item{x}{A rectangular data structure where each column is a margin and each
row is a joint realization.}

\item{p}{A rectangular data structure of probabilities.}
}
\value{
An S3 list of the \code{cma} class.
}
\description{
This function decomposes the the pure "individual" features contained in the
marginal distributions from the pure "joint" information available in the copulas.
}
\details{
Arguments \code{x} and \code{p} must have the same number of rows. The function accepts data
from the classes \code{numeric}, \code{matrix}, \code{xts} or \code{tbl}.
}
\examples{
set.seed(123)
margins <- matrix(stats::rnorm(20), ncol = 2)
probs   <- rep(1 / 20, 20)
cma_separation(x = margins, p = probs)
}
