\name{compare-methods}
\docType{methods}
\alias{compare-methods}
\alias{compare,list-method}
\title{Compare different classifiers}
\description{Compare different classifiers for the following signatures:
}
\section{Methods}{
\describe{

\item{clresultlist = "list"}{signature 1}
}
For further argument and output information, consult
\code{\link{compare}}
}
\keyword{multivariate}