% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_transformation.R
\name{copula_transformation}
\alias{copula_transformation}
\title{Copula Normalization}
\usage{
copula_transformation(x, copula)
}
\arguments{
\item{x}{A multivariate (non-tidy) dataset.}

\item{copula}{A multivariate copula.}
}
\value{
A \code{tibble}
}
\description{
This function transforms the entries of the copula to respect the boundaries
of the unit cube.
}
\details{
The copula transformation is a 4-step recipe:
    \itemize{
      \item 1. Generate new copula scenarios;
      \item 2. Normalize the copula to have a standard normal distribution;
      \item 3. Operate an affine transformation on the normalized variables;
      \item 4. Extract the copulas from the new rotated variables.
    }
}
\examples{
#
}
