% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{cube_cma_}
\alias{cube_cma_}
\title{cube_cma_
data_cube of cma}
\usage{
cube_cma_(x = mini_cma_(), omit = "libelle")
}
\arguments{
\item{x}{dataframe}

\item{omit}{regex}
}
\description{
cube_cma_
data_cube of cma
}
