% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_stats.R
\name{empirical_stats}
\alias{empirical_stats}
\title{Summary Statistics for Empirical Distributions}
\usage{
empirical_stats(x, p = NULL, level = 0.01)
}
\arguments{
\item{x}{A (non-tidy) scenario-probability distribution.}

\item{p}{A probability vector. If \code{NULL} (the default) the standard
1 over N is used.}

\item{level}{A number with the desired probability level. The default is
\code{level = 0.01}.}
}
\value{
A \code{tibble} with 2 columns and 6 rows.
}
\description{
Computes the mean, standard deviation, skewness, kurtosis, Value-at-Risk (VaR)
and Conditional Value-at-Risk CVaR) under flexible probabilities.
}
\examples{
library(ggplot2)

x <- diff(log(EuStockMarkets))

calm_market  <- panic_copula(x = x, n = 1000, panic_prob = 0.00, dist = "normal")
panic_market <- panic_copula(x = x, n = 1000, panic_prob = 0.20, dist = "normal")

# Plot Calm market stats
emp_calm <- empirical_stats(calm_market$simulation, p = calm_market$p)
ggplot(emp_calm, aes(x = name, y = value)) +
 geom_col() +
 facet_wrap(~stat, scales = "free_y") +
 scale_y_continuous(labels = scales::number_format(accuracy = 0.01))

# Plot Panic market stats
emp_panic <- empirical_stats(panic_market$simulation, p = panic_market$p)
ggplot(emp_panic, aes(x = name, y = value)) +
 geom_col() +
 facet_wrap(~stat, scales = "free_y") +
 scale_y_continuous(labels = scales::number_format(accuracy = 0.01))
}
