\name{evaluation-methods}
\docType{methods}
\alias{evaluation-methods}
\alias{evaluation,list-method}
\title{Evaluation of classifiers}
\description{Evaluate classifiers for the following signatures:}
\section{Methods}{
\describe{

\item{clresult = "list"}{signature 1}
}
For further argument and output information, consult
\code{\link{evaluation}}.
}
\keyword{multivariate}