% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{exception_}
\alias{exception_}
\title{exception
pinpoint the exceptional transactions in mainly Zanette and a few others}
\usage{
exception_(x = nosolde_())
}
\arguments{
\item{x}{path}
}
\description{
exception
pinpoint the exceptional transactions in mainly Zanette and a few others
}
