% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_moments.R
\name{ff_moments}
\alias{ff_moments}
\alias{ff_moments.default}
\alias{ff_moments.numeric}
\alias{ff_moments.matrix}
\alias{ff_moments.xts}
\alias{ff_moments.data.frame}
\alias{ff_moments.tbl}
\title{Moments under Flexible Probabilities}
\usage{
ff_moments(x, p = NULL)

\method{ff_moments}{default}(x, p = NULL)

\method{ff_moments}{numeric}(x, p = NULL)

\method{ff_moments}{matrix}(x, p)

\method{ff_moments}{xts}(x, p)

\method{ff_moments}{data.frame}(x, p)

\method{ff_moments}{tbl}(x, p)
}
\arguments{
\item{x}{A tabular (non-tidy) time series.}

\item{p}{A probability vector.}
}
\value{
A \code{list} with 2 elements: \code{mu} and \code{sigma}.
}
\description{
Computes the location and dispersion statistics for scenarios with flexible probabilities.
}
\examples{
x <- diff(log(EuStockMarkets))
p <- stats::runif(nrow(x))
p <- p /  sum(p)

ff_moments(x = matrix(x, ncol = 4), p = p)

# compare with the stadard approach
colMeans(x)
cov(x)
}
