% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_copulas.R
\name{fit_copula_clayton}
\alias{fit_copula_clayton}
\alias{fit_copula_gumbel}
\alias{fit_copula_frank}
\alias{fit_copula_t}
\alias{fit_copula_normal}
\alias{fit_copula_joe}
\title{Archimedean Copulas for Scenario Analysis}
\usage{
fit_copula_clayton(
  copula,
  method = c("mpl", "ml", "itau", "irho", "itau.mpl"),
  ...
)

fit_copula_gumbel(
  copula,
  method = c("mpl", "ml", "itau", "irho", "itau.mpl"),
  ...
)

fit_copula_frank(
  copula,
  method = c("mpl", "ml", "itau", "irho", "itau.mpl"),
  ...
)

fit_copula_t(copula, method = c("mpl", "ml", "itau", "irho", "itau.mpl"), ...)

fit_copula_normal(
  copula,
  method = c("mpl", "ml", "itau", "irho", "itau.mpl"),
  ...
)

fit_copula_joe(
  copula,
  method = c("mpl", "ml", "itau", "irho", "itau.mpl"),
  ...
)
}
\arguments{
\item{copula}{An object of the \code{cma} class.}

\item{method}{A \code{character} with the method to be used for optimization.
The default is \code{mpl}.}

\item{...}{Additional arguments to be passed to \code{\link[copula]{archmCopula}}.}
}
\value{
An S3 \code{list} of the \code{cma_copula} class.
}
\description{
Functions to build new types of copulas.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

sep <- cma_separation(x)

fit_copula_clayton(sep)
fit_copula_gumbel(sep)
}
