% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{fit_hyp}
\alias{fit_hyp}
\alias{fit_hyp.default}
\alias{fit_hyp.tbl}
\alias{fit_hyp.xts}
\alias{fit_hyp.matrix}
\title{Estimation of the Hyperbolic Distribution}
\usage{
fit_hyp(x, symmetric = FALSE)

\method{fit_hyp}{default}(x, symmetric = FALSE)

\method{fit_hyp}{tbl}(x, symmetric = FALSE)

\method{fit_hyp}{xts}(x, symmetric = FALSE)

\method{fit_hyp}{matrix}(x, symmetric = FALSE)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}

\item{symmetric}{A flag. Should the fitted distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs a maximum likelihood estimation on the Hyperbolic distribution
(univariate and multivariate). Wrappers \code{\link[ghyp]{fit.hypuv}} and
\code{\link[ghyp]{fit.hypmv}}
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

# multivariate estimation
fit_hyp(x)

# univariate estimation
fit_hyp(x[ , 4, drop = FALSE])
}
\seealso{
\code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_vg}}
\code{\link{fit_t}} \code{\link{fit_normal}}
}
