% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{fit_nig}
\alias{fit_nig}
\alias{fit_nig.default}
\alias{fit_nig.tbl}
\alias{fit_nig.xts}
\alias{fit_nig.matrix}
\title{Estimation of the Normal-Inverse Gaussian Distribution}
\usage{
fit_nig(x, symmetric = FALSE)

\method{fit_nig}{default}(x, symmetric = FALSE)

\method{fit_nig}{tbl}(x, symmetric = FALSE)

\method{fit_nig}{xts}(x, symmetric = FALSE)

\method{fit_nig}{matrix}(x, symmetric = FALSE)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}

\item{symmetric}{A \code{logical} flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs maximum likelihood estimation on the Normal-Inverse Gaussian (NIG)
distribution (univariate and multivariate). Wrappers \code{\link[ghyp]{fit.NIGuv}}
and \code{\link[ghyp]{fit.NIGmv}}.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

# multivariate estimation
fit_nig(x)

# univariate estimation
fit_nig(x[ , 4, drop = FALSE])
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_vg}}
\code{\link{fit_t}} \code{\link{fit_normal}}
}
