% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{fit_vg}
\alias{fit_vg}
\alias{fit_vg.default}
\alias{fit_vg.tbl}
\alias{fit_vg.xts}
\alias{fit_vg.matrix}
\title{Estimation of the Variance-Gamma Distribution}
\usage{
fit_vg(.invariant, .symmetric = FALSE)

\method{fit_vg}{default}(.invariant, .symmetric = FALSE)

\method{fit_vg}{tbl}(.invariant, .symmetric = FALSE)

\method{fit_vg}{xts}(.invariant, .symmetric = FALSE)

\method{fit_vg}{matrix}(.invariant, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) data structure.}

\item{.symmetric}{A \code{logical} flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs maximum likelihood estimation on the Variance-Gamma (VG)
distribution (univariate and multivariate). Wrappers \code{\link[ghyp]{fit.VGuv}}
and \code{\link[ghyp]{fit.VGmv}}.
}
\examples{
##### fit_vg #####

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

fit_vg(x[ , 3:4])     # multivariate estimation
fit_vg(x[ , 4, drop = FALSE]) # univariate estimation
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_t}}
}
