% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_moments.R
\name{fp_moments}
\alias{fp_moments}
\alias{fp_moments.default}
\alias{fp_moments.numeric}
\alias{fp_moments.matrix}
\alias{fp_moments.xts}
\alias{fp_moments.data.frame}
\alias{fp_moments.tbl}
\title{Moments under Flexible Probabilities}
\usage{
fp_moments(x, p = NULL)

\method{fp_moments}{default}(x, p = NULL)

\method{fp_moments}{numeric}(x, p = NULL)

\method{fp_moments}{matrix}(x, p = NULL)

\method{fp_moments}{xts}(x, p = NULL)

\method{fp_moments}{data.frame}(x, p = NULL)

\method{fp_moments}{tbl}(x, p = NULL)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}

\item{p}{A probability vector.}
}
\value{
A \code{list} with 2 elements: \code{mu} and \code{sigma}.
}
\description{
Computes the location and dispersion statistics under flexible probabilities.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)
colnames(x) <- colnames(EuStockMarkets)
p <- stats::runif(nrow(x))
p <- p / sum(p)

fp_moments(x = x, p = p)

# compare with the standard approach
colMeans(x)
cov(x)
}
