% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{fwrite_long_}
\alias{fwrite_long_}
\title{fwrite_long_
fast write of cube file}
\usage{
fwrite_long_(x = mini_long_(), y = "cma_long.csv")
}
\arguments{
\item{x}{dataframe}

\item{y}{filename}
}
\description{
fwrite_long_
fast write of cube file
}
