% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{fwrite_raw_}
\alias{fwrite_raw_}
\title{fwrite_raw_
fast write of csv file}
\usage{
fwrite_raw_(x = read_all_(), y = "cma_raw.csv")
}
\arguments{
\item{x}{dataframe}

\item{y}{filename}
}
\description{
fwrite_raw_
fast write of csv file
}
