% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_copulas}
\alias{generate_copulas}
\title{Generate Copulas}
\usage{
generate_copulas(model, n)
}
\arguments{
\item{model}{An object of the \code{cma_copula} class.}

\item{n}{An \code{integer} with the number of samples to be generated.}
}
\value{
An S3 \code{list} of the class \code{bla bla bla}.
}
\description{
This function takes a fitted object from the \code{cma_copula} class and generate
new scenarios for the copula
}
\examples{
#
}
