\name{internals}
\alias{plotprob}
\alias{roundvector}
\alias{rowswaps}
\alias{ROCinternal}
\alias{bklr}
\alias{bkreg}
\alias{bklr.predict}
\alias{care.exp}
\alias{care.dev}
\alias{mklr}
\alias{mkreg}
\alias{mklr.predict}
\alias{my.care.exp}
\alias{characterplot}
\alias{safeexp}
\title{Internal functions}
\description{Not intended to be called directly by the user.}
\keyword{multivariate}
