\name{join-methods}
\docType{methods}
\alias{join-methods}
\alias{join,list-method}
\title{Combine list elements returned by the method classification}
\description{
 The list of objects of class \code{\link{cloutput}} can be unified
 into one object for the following signatures:
}
\section{Methods}{
\describe{
\item{cloutputlist = "list"}{signature 1}
}
For further argument and output information, consult
\code{\link{join}}.
}
\keyword{multivariate}
