\name{ldaCMA-methods}
\docType{methods}
\alias{ldaCMA-methods}
\alias{ldaCMA,matrix,numeric,missing-method}
\alias{ldaCMA,matrix,factor,missing-method}
\alias{ldaCMA,data.frame,missing,formula-method}
\alias{ldaCMA,ExpressionSet,character,missing-method}
\title{Linear Discriminant Analysis}
\description{
 Performs a linear discriminant analysis for the following signatures:
}
\section{Methods}{
\describe{

\item{X = "matrix", y = "numeric", f = "missing"}{signature 1}

\item{X = "matrix", y = "factor", f = "missing"}{signature 2}

\item{X = "data.frame", y = "missing", f = "formula"}{signature 3}

\item{X = "ExpressionSet", y = "character", f = "missing"}{signature 4}
}
For further argument and output information, consult
\code{\link{ldaCMA}}.
}
\keyword{multivariate}

