% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{long_cma_}
\alias{long_cma_}
\title{long_cma_
data_cube of cma}
\usage{
long_cma_(x = fread_raw_(), omit = "Xlibelle")
}
\arguments{
\item{x}{dataframe}

\item{omit}{regex}
}
\description{
long_cma_
data_cube of cma
}
