% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_moments_normal.R
\name{match_moments_normal}
\alias{match_moments_normal}
\title{Simulations with Exact Means and Covariances}
\usage{
match_moments_normal(M, S, J)
}
\arguments{
\item{M}{A \code{1xN} matrix with the location parameter of each invariant (asset).}

\item{S}{A \code{NxN} matrix with the dispersion matrix of the invariants (assets).}

\item{J}{A numeric scalar with the desired number of scenarios.}
}
\value{
A J x N matrix with the simulated time-series.
}
\description{
Generates scenarios from elliptical distributions in which the
sample moments match the populational moments.
}
\examples{
rets  <- diff(log(EuStockMarkets))
mu    <- colMeans(rets)
sigma <- stats::cov(rets)
match_moments_normal(M = mu, S = sigma, J = 10)
}
