% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_matching.R
\name{match_normal}
\alias{match_normal}
\title{Moment Matching for the Normal Distribution}
\usage{
match_normal(mu, sigma, n)
}
\arguments{
\item{mu}{A \code{double} with the location parameter.}

\item{sigma}{A variance-covariance \code{matrix} the dispersion parameter.}

\item{n}{An \code{integer} with the number of scenarios to be generated.}
}
\value{
#TODO
}
\description{
Generates scenarios from the normal distribution that exactly matches
the first two moments.
}
\examples{
x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
match_normal(mu, sigma, 10)

# correlation structure is similar
cor(x)
cor(match_normal(mu, sigma, 10000))

# as the location parameters
mu
colMeans(match_normal(mu, sigma, 10000))
}
\seealso{
\code{\link{match_t}}
}
