% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_matching.R
\name{match_t}
\alias{match_t}
\title{Moment Matching for the t Distribution}
\usage{
match_t(mu, sigma, nu, n, stochastic = FALSE)
}
\arguments{
\item{mu}{A \code{double} with the location parameter.}

\item{sigma}{A variance-covariance \code{matrix} the dispersion parameter.}

\item{nu}{An \code{integer} with the degrees of freedom.}

\item{n}{An \code{integer} with the number of scenarios to be generated.}

\item{stochastic}{A \code{logical} flag indicating whether (or not) the stochastic
representation should be used. The default if \code{TRUE}.}
}
\value{
A \code{tibble} in the wide format.
}
\description{
Generates scenarios from the Student-t distribution that
exactly matches the first two moments.
}
\examples{
x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
nu    <- 5
match_t(mu, sigma, nu, 10)

cor(x)
cor(match_t(mu, sigma, 5, 100000))

mu
colMeans(match_t(mu, sigma, nu, 100000))
}
\seealso{
\code{\link{match_normal}}
}
