% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{mini_cma_}
\alias{mini_cma_}
\title{mini_cma_
mini_tibble of selected columns}
\usage{
mini_cma_(
  x = group_cma_(),
  omit = "^date|^tab|^code|^operation|^valeur|^bank|^comptable"
)
}
\arguments{
\item{x}{dataframe}

\item{omit}{regex}
}
\description{
mini_cma_
mini_tibble of selected columns
}
