% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{mini_long_}
\alias{mini_long_}
\title{mini_long_
mini_tibble long for 2500 rows of selected columns}
\usage{
mini_long_(
  x = long_cma_(),
  omit = "^Xdc$|period|group|^date|^tab|^code|^operation|^valeur|^bank|^comptable"
)
}
\arguments{
\item{x}{dataframe}

\item{omit}{regex}
}
\description{
mini_long_
mini_tibble long for 2500 rows of selected columns
}
