% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{new_cma_fit}
\alias{new_cma_fit}
\alias{obj_print_data.cma_fit}
\title{Internal function to create a prayer fit class.}
\usage{
new_cma_fit(x, ...)

\method{obj_print_data}{cma_fit}(x, ...)
}
\arguments{
\item{x}{A fitted object}

\item{...}{Any arguments to be passed as attributes}
}
\value{
A \code{list} of the \code{prayer_fit} class.
}
\description{
Internal function to create a prayer fit class.
}
\keyword{internal}
