% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-cma.R
\name{new_cma_separation}
\alias{new_cma_separation}
\title{Object created to make printing prettier.}
\usage{
new_cma_separation(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments.}
}
\value{
An object of the cma class.
}
\description{
Object created to make printing prettier.
}
\keyword{internal}
