% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panic_copula.R
\name{panic_copula}
\alias{panic_copula}
\title{Panic-Copula for Stress-Testing}
\usage{
panic_copula(
  x,
  n = 10000,
  panic_cor = 0.99,
  panic_prob = 0.02,
  dist = c("normal", "t")
)
}
\arguments{
\item{x}{A rectangular (non-tidy) data structure.}

\item{n}{An \code{integer} with the number of scenarios to be generated.}

\item{panic_cor}{A numeric value for the correlation in panic markets.}

\item{panic_prob}{A numeric value with the probability in which panic markets
can be triggered.}

\item{dist}{A \code{character} with \code{normal} or \code{t}. The default is
\code{normal}}
}
\value{
An object of the \code{panic} class.
}
\description{
This function generates a large dimensional panic copula that can be further
used for stress-testing and panic-aware portfolio optimization.
}
\examples{
x <- diff(log(EuStockMarkets))
panic_copula(x = x, n = 20, panic_cor = 0.99, panic_prob = 0.02)
}
