% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panic_distribution.R
\name{plot_panic_distribution}
\alias{plot_panic_distribution}
\title{Visualize a Panic Distribution}
\usage{
plot_panic_distribution(pnl, p = NULL, breaks = 200)
}
\arguments{
\item{pnl}{An univariante time series with the PnL marginal distribution.}

\item{p}{A \code{double} vector of probabilities. If \code{NULL}, `p` is automatically
set to 1 over n.}

\item{breaks}{An \code{integer} with the number of break point to be used to
plot the marginal panic distribution.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualize a Panic Distribution
}
\examples{
x <- diff(log(EuStockMarkets))
x <- panic_copula(x = x, n = 10000, panic_prob = 0.02)
w <- rep(0.25, 4)
pnl <- x$simulation \%*\% w
plot_panic_distribution(pnl = pnl, p = x$p)
}
