% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{read.stockholm.alignment}
\alias{read.stockholm.alignment}
\title{Reads an sth file}
\usage{
read.stockholm.alignment(file = "Y://PFAM2/data/sth/seed/PF00013_v27_seed.sth")
}
\arguments{
\item{file}{sth file name}
}
\value{
a dataframe with name and seq columnns
}
\description{
\code{read.stockholm.alignment} reads a stockholm file format
Only lines which do not start with a # are considered
}
