% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{read_sheet_}
\alias{read_sheet_}
\title{read_sheet_
read one sheet of data}
\usage{
read_sheet_(x = sheets_(y)[1], y = path_("CMCC2019.xlsx"))
}
\arguments{
\item{x}{sheet name}

\item{y}{file path}
}
\description{
read_sheet_
read one sheet of data
}
