% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{read_year_}
\alias{read_year_}
\title{read_year_
read one year of data}
\usage{
read_year_(
  year = 2021,
  x = sheets_(y),
  y = path_(paste0("CMCC", year, ".xlsx"))
)
}
\arguments{
\item{year}{integer}

\item{x}{vector of sheet names}

\item{y}{file path}
}
\description{
read_year_
read one year of data
}
