% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_pooling.R, R/ffp_moments.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{entropy_pooling}
\alias{ffp_moments}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ffp}{\code{\link[ffp]{entropy_pooling}}, \code{\link[ffp]{ffp_moments}}}
}}

