% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize}
\alias{sanitize}
\alias{sanitize.default}
\alias{sanitize.matrix}
\alias{sanitize.ts}
\alias{sanitize.xts}
\alias{sanitize.zoo}
\alias{sanitize.data.frame}
\alias{sanitize.tbl_df}
\title{Generic that internally checks input arguments}
\usage{
sanitize(x)

\method{sanitize}{default}(x)

\method{sanitize}{matrix}(x)

\method{sanitize}{ts}(x)

\method{sanitize}{xts}(x)

\method{sanitize}{zoo}(x)

\method{sanitize}{data.frame}(x)

\method{sanitize}{tbl_df}(x)
}
\description{
Generic that internally checks input arguments
}
\keyword{internal}
