% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{sheets_}
\alias{sheets_}
\title{sheets_
retrieve sheet names in excel file}
\usage{
sheets_(x = path_(y), y = "CMCC2021.xlsx")
}
\arguments{
\item{x}{file path}

\item{y}{last filename}
}
\description{
sheets_
retrieve sheet names in excel file
}
