% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_normal}
\alias{simulate_normal}
\title{Moment Matching Algorithm for the Normal Distribution}
\usage{
simulate_normal(mu, sigma, n)
}
\arguments{
\item{mu}{A \code{double} with the location parameter.}

\item{sigma}{A variance-covariance \code{matrix} the dispersion parameter.}

\item{n}{A \code{integer} with the number of simulations to be conducted.}
}
\value{
#TODO
}
\description{
This function generates scenarios from the normal distribution that match
the sample counterparts of given dataset.
}
\examples{
x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
simulate_normal(mu, sigma, 10)

cor(x)
cor(simulate_normal(mu, sigma, 10000))

mu
colMeans(simulate_normal(mu, sigma, 10000))
}
