\name{varseloutput-class}
\docType{class}
\alias{varseloutput-class}
\alias{varseloutput}
\title{"varseloutput"}
\description{An object returned by the functions described in \code{\link{filter}},
             usually not created directly by the user.}
\section{Slots}{
	 \describe{
    \item{\code{varsel}:}{\code{numeric} vector of variable importance measures, e.g.
                          absolute of genewise statistics.}
  }
}
\section{Methods}{No methods are currently defined.}

\author{Martin Slawski \email{ms@cs.uni-sb.de}

        Anne-Laure Boulesteix \email{boulesteix@ibe.med.uni-muenchen.de}}

\seealso{\code{\link{filter}}, \code{\link{clvarseloutput}}}
\keyword{multivariate}
