\name{wmc-methods}
\docType{methods}
\alias{wmc-methods}
\alias{wmc,matrix,numeric,numeric-method}

\title{General method for tuning / selection bias correction based on a
  matrix of subsampling fold errors.}
\description{
  Perform tuning / selection bias correction for a matrix of subsampling fold errors
  for the following signature:
}
\section{Methods}{
\describe{

  \item{mcr.m="matrix",n.tr="numeric",n.ts="numeric"}{signature 1}
}

For further argument and output information, consult
\code{\link{wmc}}.}
\keyword{tuning bias, selection bias, corrected misclassification rate}