% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMA.R
\docType{package}
\name{CMA}
\alias{CMA}
\title{CMA: Copula-Marginal Algorithm}
\description{
The Copula-Marginal Algorithm (CMA) provides a tool to estimate
fully flexible copulas. CMA provides a simple two-step algorithm to decompose
a multivariate distribution in copulas and marginals (of any kind).
This flexibility follows from the fact that, unlike traditional approaches to
copulas, CMA does not require the explicit computation of marginal CDF and their
inverses. As a result, CMA can generate scenarios for many more copulas than the
few parametric families used in the traditional approach.
}
