% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_combination.R
\name{cma_combination}
\alias{cma_combination}
\title{CMA combination Step}
\usage{
cma_combination(x, u, U)
}
\arguments{
\item{x}{A rectangular (non-tidy) data structure with the sorted marginals
(ascending order).#' .}

\item{u}{A rectangular (non-tidy) data structure with the CDF's.}

\item{U}{A rectangular (non-tidy) data structure with the copulas.}
}
\value{
A \code{tbl} with the joint scenarios.
}
\description{
This function glues back any arbitrary copula to marginal distributions
of any kind.
}
\examples{
set.seed(123)

margins <- matrix(stats::rnorm(20), ncol = 2)
colnames(margins) <- c("a", "b")
probs   <- rep(1 / 25, 25)

# separate
sep <- cma_separation(x = margins, p = probs)
# combinate
comb <- cma_combination(x = attributes(sep)$ord_margin, u = attributes(sep)$cdf, U = sep$copula)

# The result is identical
stopifnot(all.equal(margins, comb))
}
