% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_combination.R
\name{cma_combination}
\alias{cma_combination}
\title{CMA combination Step}
\usage{
cma_combination(x, cdf, copula)
}
\arguments{
\item{x}{A rectangular (non-tidy) data structure with the sorted marginals
(ascending order).}

\item{cdf}{A rectangular (non-tidy) data structure with the CDF's.}

\item{copula}{A rectangular (non-tidy) data structure with the copulas.}
}
\value{
A \code{tbl} with the joint scenarios.
}
\description{
Second CMA step: "glue" copulas and marginals into the same data structure.
}
\examples{
set.seed(123)

margins <- matrix(stats::rnorm(20), ncol = 2)
colnames(margins) <- c("a", "b")
probs   <- rep(1 / 10, 10)

# separate
sep <- cma_separation(x = margins, p = probs)
# combinate
comb <- cma_combination(margins, sep$cdf, sep$copula)

# The result is identical
margins
comb
stopifnot(all.equal(margins[ , "a"], comb$a))
stopifnot(all.equal(margins[ , "b"], comb$b))
}
\seealso{
\code{\link{cma_separation}}
}
