% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_separation.R
\name{cma_separation}
\alias{cma_separation}
\alias{cma_separation.default}
\alias{cma_separation.matrix}
\alias{cma_separation.xts}
\alias{cma_separation.data.frame}
\alias{cma_separation.tbl}
\title{CMA Separation Step}
\usage{
cma_separation(x, p = NULL)

\method{cma_separation}{default}(x, p = NULL)

\method{cma_separation}{matrix}(x, p = NULL)

\method{cma_separation}{xts}(x, p = NULL)

\method{cma_separation}{data.frame}(x, p = NULL)

\method{cma_separation}{tbl}(x, p = NULL)
}
\arguments{
\item{x}{A rectangular (non-tidy) data structure.}

\item{p}{A probability vector. If \code{NULL} (the default) the standard
1 over N is used.}
}
\value{
An S3 list of the \code{cma} class that contains two elements:
\code{marginal} and \code{copula}.
}
\description{
First CMA step: decomposes the the pure "individual" features of the marginal
from the pure "joint" information available in the copulas.
}
\details{
Arguments \code{x} and \code{p} must have the same size.
}
\examples{
set.seed(123)
margins <- matrix(stats::rnorm(30), ncol = 3)
cma <- cma_separation(x = margins)
cma

# Access the elements with `$`
cma$copula
cma$marginal
cma$cdf
}
\seealso{
\code{\link{cma_combination}}
}
