% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_transformation.R
\name{copula_transformation}
\alias{copula_transformation}
\title{Copula Normalization}
\usage{
copula_transformation(x, copula)
}
\arguments{
\item{x}{A multivariate (non-tidy) dataset.}

\item{copula}{A multivariate copula.}
}
\value{
A \code{tibble}
}
\description{
This function transforms the entries an arbitrary copula so they can be compatible
with the first two moments of a timeseries.
}
\details{
Under the hood, the copula transformation is a 4-step recipe:
    \itemize{
      \item 1. Normalize the copula to have a standard normal distribution;
      \item 2. Operate an affine transformation on the normalized variables;
      \item 3. Do the CMA Separation step on the new rotated variables;
      \item 4. Extract the copula component that results from step 3.
    }
}
\examples{
x       <- matrix(diff(log(EuStockMarkets)), ncol = 4)

sep     <- cma_separation(x)
clayton <- fit_copula_clayton(sep)
gen     <- generate_copulas(clayton, 10)

copula_transformation(x, gen)
}
