% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{fit_ghd}
\alias{fit_ghd}
\alias{fit_ghd.default}
\alias{fit_ghd.tbl}
\alias{fit_ghd.xts}
\alias{fit_ghd.matrix}
\title{Estimation of the Generalized Hyperbolic Distribution}
\usage{
fit_ghd(x, symmetric = FALSE)

\method{fit_ghd}{default}(x, symmetric = FALSE)

\method{fit_ghd}{tbl}(x, symmetric = FALSE)

\method{fit_ghd}{xts}(x, symmetric = FALSE)

\method{fit_ghd}{matrix}(x, symmetric = FALSE)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}

\item{symmetric}{A \code{logical} flag. Should the fitted distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs the estimation on the Generalized Hyperbolic distribution
(univariate and multivariate). Wrappers \code{\link[ghyp]{fit.ghypuv}} and
\code{\link[ghyp]{fit.ghypmv}}.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

# multivariate estimation
fit_ghd(x)

# univariate estimation
fit_ghd(x[ , 3, drop = FALSE])
}
\seealso{
\code{\link{fit_hyp}} \code{\link{fit_nig}} \code{\link{fit_vg}}
\code{\link{fit_t}} \code{\link{fit_normal}}
}
