% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{fit_t}
\alias{fit_t}
\alias{fit_t.default}
\alias{fit_t.tbl}
\alias{fit_t.xts}
\alias{fit_t.matrix}
\title{Estimation of the Student-t Distribution}
\usage{
fit_t(x, symmetric = FALSE)

\method{fit_t}{default}(x, symmetric = FALSE)

\method{fit_t}{tbl}(x, symmetric = FALSE)

\method{fit_t}{xts}(x, symmetric = FALSE)

\method{fit_t}{matrix}(x, symmetric = FALSE)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}

\item{symmetric}{A \code{logical} flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs maximum likelihood estimation on the Student-t distribution
(univariate and multivariate). Wrappers \code{\link[ghyp]{fit.tuv}}
and \code{\link[ghyp]{fit.tmv}}.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

# multivariate estimation
fit_t(x)

# univariate estimation
fit_t(x[ , 4, drop = FALSE])
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_nig}}
\code{\link{fit_vg}} \code{\link{fit_normal}}
}
