% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margins.R
\name{fit_vg}
\alias{fit_vg}
\alias{fit_vg.default}
\alias{fit_vg.tbl}
\alias{fit_vg.xts}
\alias{fit_vg.matrix}
\title{Estimation of the Variance-Gamma Distribution}
\usage{
fit_vg(x, symmetric = FALSE)

\method{fit_vg}{default}(x, symmetric = FALSE)

\method{fit_vg}{tbl}(x, symmetric = FALSE)

\method{fit_vg}{xts}(x, symmetric = FALSE)

\method{fit_vg}{matrix}(x, symmetric = FALSE)
}
\arguments{
\item{x}{A tabular (non-tidy) data structure.}

\item{symmetric}{A \code{logical} flag. Should the fitted distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} of the the class \code{cma_fit} with \code{21} components.
}
\description{
Performs the estimation on the Variance-Gamma (VG)
distribution (univariate and multivariate). Wrappers \code{\link[ghyp]{fit.VGuv}}
and \code{\link[ghyp]{fit.VGmv}}.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)

# multivariate estimation
fit_vg(x[ , 3:4])

# univariate estimation
fit_vg(x[ , 4, drop = FALSE])
}
\seealso{
\code{\link{fit_ghd}} \code{\link{fit_hyp}} \code{\link{fit_nig}}
\code{\link{fit_t}} \code{\link{fit_normal}}
}
