% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_copulas}
\alias{generate_copulas}
\title{Generate Copulas}
\usage{
generate_copulas(model, n)
}
\arguments{
\item{model}{An object of the \code{cma_copula} class.}

\item{n}{An \code{integer} with the number of samples to be generated.}
}
\value{
An S3 \code{list} of the class \code{bla bla bla}.
}
\description{
This function takes a fitted object from the \code{cma_copula} class and generate
new scenarios for the copula
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)
colnames(x) <- c("a", "b", "c", "d")

# fit normal copula
sep  <- cma_separation(x)
normcop <- fit_copula_normal(sep)

generate_copulas(normcop, 10)
}
