% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma-vctrs.R
\name{new_cma_separation}
\alias{new_cma_separation}
\alias{new_cma_to_print}
\title{Object created to make printing prettier.}
\usage{
new_cma_separation(x, ...)

new_cma_to_print(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments.}
}
\value{
An objecto of the cma class.
}
\description{
Object created to make printing prettier.
}
\keyword{internal}
