% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panic_copula.R
\name{panic_copula}
\alias{panic_copula}
\title{Panic-Copula for Stress-Testing}
\usage{
panic_copula(
  N = 2,
  J = 50000,
  calm_cor = 0.3,
  panic_cor = 0.99,
  panic_prob = 0.02,
  sigma = 0.2
)
}
\arguments{
\item{N}{A numeric value with the number of assets.}

\item{J}{A numeric value with the number of scenarios to be generated.}

\item{calm_cor}{A numeric value for the correlation in calm markets.}

\item{panic_cor}{A numeric value for the correlation in panic markets.}

\item{panic_prob}{A numeric value with the probability in which panic markets
can be triggered.}

\item{sigma}{The unconditional volatility of the joint-distribution.}
}
\value{
An object with \code{N} columns and \code{J} rows.
}
\description{
This function generates a large dimensional panic copula that can be further
used for stress-testing and panic-aware portfolio optimization.
}
\examples{
panic_copula(N = 2, J = 10, calm_cor = 0.3, panic_cor = 0.99, panic_prob = 0.02, sigma = 0.2)
}
